refactor: Reorganize CallDetailsModel
