+++ Aircall/Aircall/Scenes/CallDetails/Model/CallDetailsModel+FetchRequest.swift
@@ -1,15 +1,14 @@
 //
-//  CallListModel+FetchRequest.swift
+//  CallDetailsModel+FetchRequest.swift
 //  Aircall
 //
 //  Created by Yanis SIDAHMED on 16/07/2019.
 //  Copyright © 2019 Yanis SG. All rights reserved.
 //
 
-import Foundation
 import CoreData
 
-extension CallListModel {
+extension CallDetailsModel {
 
 	struct FetchRequest {
 
@@ -22,20 +21,28 @@ extension CallListModel {
 		  - managedObjectModel: Managed object to model to search in.
 
 		- Returns: Fetch request ready to be used.
-		
+
 		*/
 		static func requestBuilder(
-			parameter: CallListModel.Parameter,
+			parameter: CallDetailsModel.Parameter,
 			managedObjectModel: NSManagedObjectModel
 			) -> NSFetchRequest<CDCall> {
+
 			guard let fetchRequest = managedObjectModel.fetchRequestFromTemplate(
-				withName: parameter.scope.key,
+				withName: CDCall.Scope.all.key,
 				substitutionVariables: [:]
 				) as? NSFetchRequest<CDCall> else {
 					fatalError("fetchRequest type is not NSFetchRequest<CDCall> as expected.")
 			}
-			let sortDescriptor = NSSortDescriptor(key: "created", ascending: parameter.ascending)
+			let sortDescriptor = NSSortDescriptor(key: "id", ascending: true)
 			fetchRequest.sortDescriptors = [sortDescriptor]
+
+			let predicates: [NSPredicate] = [
+				fetchRequest.predicate,
+				NSPredicate(format: "id = %i", parameter.callId)
+				].compactMap {$0}
+			fetchRequest.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: predicates)
+
 			return fetchRequest
 		}
